+++ src/App.js
@@ -3,11 +3,16 @@ import TODOListMainView from "./components/TODOList/TODOListMainView";
 import "./App.css";
 
 function App() {
-  return (
-    <>
-      <SideBar></SideBar>
-    </>
-  );
+    return (
+        <div className="overallAppView">
+            <div className="sideBarView">
+                <SideBar />
+            </div>
+            <div className="dashBoardView">
+                <TODOListMainView/>
+            </div>
+        </div>
+    );
 }
 
 export default App;
